# docs/openapi/eomp-openapi-spec.yaml
openapi: 3.0.3
info:
  title: EOMP API
  version: 0.1.0
  description: |
    Runtime operation baseline for the EOMP gateway and its downstream services.
    All 102 registered `/api/v1` operations are represented. Generic write bodies
    mark endpoints whose domain schemas still need richer component definitions.
  x-coverage-status: complete-operation-inventory
  x-runtime-operation-count: 102
  x-documented-operation-count: 102
servers:
  - url: http://localhost:8080
    description: Local gateway
security:
  - BearerAuth: []
paths:
...
  /api/v1/workflows/instances:
    get: {tags: [Workflow], operationId: listWorkflowInstances, summary: List workflow instances (employees are requester-scoped), responses: {'200': {$ref: '#/components/responses/Success'}}}
    post: {tags: [Workflow], operationId: startWorkflow, summary: Atomically start a workflow with its configured role-pool approval and initial log, requestBody: {$ref: '#/components/requestBodies/JsonBody'}, responses: {'201': {$ref: '#/components/responses/Created'}}}
...
components:
  securitySchemes:
    BearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  responses:
    Success: {description: Successful response}
    Created: {description: Resource created}
...
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: {type: object, additionalProperties: true}
...